third_party/WebKit/LayoutTests/css3/device-adapt/viewport-insert-rule-before-expected.txt

ALERT: viewport size 500x550 scale 0.64 with limits [0.64, -1] and userScalable true
